logTimeStamp/Jia: 1. ProgressBar抽出成共用格式Module置於common，並修改Modules的TimeStamp, ProgressBar
