+++ src/assets/styles/Variables.scss
@@ -19,3 +19,9 @@ $colorPink: #f8c7cc;
 $colorOrange: #e9724c;
 $colorDeepBlue: #253d5b;
 $colorCitron: #c3bf6d;
+$colorGray: #f0f0f0;
+$colorTeaGreen: #c9f2c7;
+$colorLightGreen: #aceca1;
+$colorOlivine: #96be8c;
+$colorAsparagus: #629460;
+$colorDarkGreen: #243119;
