+++ src/components/Modules/Countdown/InputData/InputData.scss
@@ -32,19 +32,6 @@
     font-size: 5rem;
   }
 
-  &__btn {
-    font-size: 3rem;
-    border-radius: 20px;
-    padding: 10px 20px;
-    background-color: $colorGreen;
-    color: $colorWhite;
-    font-weight: bold;
-    border: 10px solid $colorWhite;
-    width: 100%;
-    max-width: 500px;
-    font-size: 4rem;
-  }
-
   &__currentTask {
     text-align: center;
     font-size: 3rem;
