+++ src/components/Modules/Countdown/ProgressBarModal/ProgressBarModal.scss
@@ -40,101 +40,11 @@
     }
   }
 
-  &__bar {
-    display: flex;
-    flex-flow: row nowrap;
-    align-items: start;
-    justify-content: center;
-    width: 480px;
-    max-width: 80%;
-    @include phone {
-      width: 250px;
-    }
-  }
-
   &__title {
     font-size: 2rem;
     font-weight: bold;
     @include phone {
       font-size: 1.5rem;
     }
-  }
-
-  &__segment {
-    display: flex;
-    font-size: 0;
-    height: 30px;
-    overflow: hidden;
-
-    // transition: width 0.3s ease;
-    &--focus {
-      background: darken(#4caf50, 10%);
-    }
-    &--distract {
-      background: darken(#f44336, 10%);
-    }
-  }
-
-  
-}
-
-.scheduledProgressBar {
-  &__bar {
-    display: flex;
-    flex-flow: row nowrap;
-    align-items: start;
-    justify-content: center;
-    width: 480px;
-    max-width: 80%;
-    @include phone {
-      width: 250px;
-    }
-  }
-
-  &__title {
-    font-size: 2rem;
-    font-weight: bold;
-    @include phone {
-      font-size: 1.5rem;
-    }
-  }
-
-  &__segment {
-    display: flex;
-    font-size: 0;
-    height: 30px;
-    overflow: hidden;
-
-    // transition: width 0.3s ease;
-    &--focus {
-      background: darken(#4caf50, 10%);
-    }
-    &--distract {
-      background: darken(#f44336, 10%);
-    }
-  }
-}
-
-.timeStamp {
-  display: flex;
-  flex-flow: row nowrap;
-  font-size: 2rem;
-  width: 100%;
-  padding: 0px 20px;
-  justify-content: space-between;
-  @include phone {
-    font-size: 1.5rem;
-    padding: 0px 10px;
-  }
-  &__start {
-    justify-content: start;
-  }
-
-  &__end {
-    justify-content: end;
-  }
-
-  &__block {
-    display: none;
   }
 }
