+++ src/components/Modules/Countdown/ProgressBarModal/ProgressBarModal.tsx
@@ -7,41 +7,9 @@ import { useSegmentProgressBar } from '@/hooks/useSegmentProgressBar';
 import { useScheduledSegments } from '@/hooks/useScheduledSegment';
 import { DEFAULT_WORKTIME, DEFAULT_BREAKTIME } from '@/constants/storage';
 import Button, { ButtonVariant } from '@/components/common/Button/Button';
-import type { TimeStampProp } from '@/types/taskTypes';
+import ProgressBar from '@/components/common/ProgressBar/ProgressBar';
+import TimeStamp from '@/components/common/TimeStamp/TimeStamp';
 import './ProgressBarModal.scss';
-
-const TimeStamp = ({ start, end, block }: TimeStampProp) => {
-  return (
-    <div className="timeStamp">
-      <div className="timeStamp__start">
-        {new Date(start).toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}
-      </div>
-      <div className="timeStamp__end">
-        {new Date(end).toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}
-      </div>
-      <div className="timeStamp__block">{block}</div>
-    </div>
-  );
-};
-const ProgressBar = ({
-  segments,
-  block,
-  stateMap,
-}: {
-  segments: any[];
-  block: string;
-  stateMap: Record<string, string>;
-}) => (
-  <div className={`${block}__bar`}>
-    {segments.map((seg, index) => (
-      <div
-        key={index}
-        style={{ width: `${seg.percent}%` }}
-        className={`${block}__segment ${block}__segment--${stateMap[seg.state] || seg.state}`}
-      />
-    ))}
-  </div>
-);
 
 const ProgressBarModal = () => {
   const { pageViewLog = [] } = useAtomValue(currentTaskAtom);
@@ -74,10 +42,10 @@ const ProgressBarModal = () => {
           stateMap={{ focus: 'focus', distract: 'distract', unknown: 'unknown' }}
         />
 
-        <div className="scheduledProgressBar__title">Planned Progress Bar</div>
+        <div className="progressBar__title">Planned Progress Bar</div>
         <ProgressBar
           segments={scheduledSegments}
-          block="scheduledProgressBar"
+          block="progressBar"
           stateMap={{ work: 'focus', break: 'distract' }} //work跟break的狀態映射
         />
         <div className="btn btn__wrapper">
