+++ src/components/common/ProgressBar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProgressBar';
