+++ src/components/common/TimeStamp/TimeStamp.scss
@@ -1,0 +1,23 @@
+.timeStamp {
+  display: flex;
+  flex-flow: row nowrap;
+  font-size: 2rem;
+  width: 100%;
+  padding: 0px 20px;
+  justify-content: space-between;
+  @include phone {
+    font-size: 1.5rem;
+    padding: 0px 10px;
+  }
+  &__start {
+    justify-content: start;
+  }
+
+  &__end {
+    justify-content: end;
+  }
+
+  &__block {
+    display: none;
+  }
+}
